# openapi.yaml
openapi: 3.1.0
info:
  title: Scrutineer Skill API
  version: 0.4.0
  description: |
    A narrow HTTP surface that claude-code skills use while they run. The
    goal is enough to let a skill orchestrate other skills and read prior
    scan outputs without pushing that logic into scrutineer's Go code.

    Every scrutineer skill scan is issued a bearer token that is staged into
    the workspace context.json. Skills authenticate with that token. A token
    is only valid while its originating scan is running, and only grants
    access to the repository the scan was issued against.

    The `/v1/*` paths are an unauthenticated, localhost-only export surface
    intended for external scripts (jq, cron, CI dumps). They emit JSONL and
    rely on the same host-header check as the browser UI; see threatmodel.md.
servers:
  - url: http://127.0.0.1:8080/api
    description: Local scrutineer server
security:
  - bearerAuth: []
paths:
...
  /findings/{id}/notes:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      summary: List notes on a finding
      operationId: listFindingNotes
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FindingNote' }
    post:
      summary: Append a note to a finding
      operationId: addFindingNote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
                by:   { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingNote' }
...
  /findings/{id}/communications:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      summary: List logged communications for a finding
      operationId: listFindingCommunications
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FindingCommunication' }
    post:
      summary: Log a communication (sent or received) about a finding
      operationId: addFindingCommunication
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FindingCommunication' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingCommunication' }
  /findings/{id}/references:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      summary: List references attached to a finding
      operationId: listFindingReferences
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FindingReference' }
    post:
      summary: Attach an external reference to a finding
      operationId: addFindingReference
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FindingReference' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingReference' }
  /findings/{id}/labels:
    put:
      summary: Replace the set of labels on a finding
      operationId: setFindingLabels
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [labels]
              properties:
                labels:
                  type: array
                  items: { type: string }
                  description: |
                    Label names, replacing the finding's current set. Missing
                    labels are created on demand. Blank and whitespace-only
                    names are trimmed and ignored, so a request whose names are
                    all blank clears every label just as `[]` does. The field
                    itself is required: omitting it or sending null is rejected
                    rather than treated as a clear-all.
      responses:
        '204': { description: Updated }
        '400':
          description: Body is not valid JSON, or has no labels array
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    FindingNote:
      type: object
      required: [id, finding_id, body]
      properties:
        id:         { type: integer, format: int64 }
        finding_id: { type: integer, format: int64 }
        body:       { type: string }
        by:         { type: string }
        created_at: { type: string, format: date-time }
...
    FindingCommunication:
      type: object
      required: [channel, direction, actor]
      properties:
        id:           { type: integer, format: int64 }
        finding_id:   { type: integer, format: int64 }
        channel:      { type: string, enum: [email, ghsa, issue, pr, direct, registry] }
        direction:    { type: string, enum: [outbound, inbound] }
        actor:        { type: string }
        body:         { type: string }
        offered_help: { type: string, enum: [pr, funding, adoption, "none", ""] }
        at:           { type: string, format: date-time }
        created_at:   { type: string, format: date-time }
    FindingReference:
      type: object
      required: [url]
      properties:
        id:         { type: integer, format: int64 }
        finding_id: { type: integer, format: int64 }
        url:        { type: string, format: uri }
        tags:       { type: string, description: Comma-joined tags (issue, pr, cve, ghsa, patch, advisory, discussion, article). }
        summary:    { type: string }
        created_at: { type: string, format: date-time }
...